@@ -3,13 +3,8 @@ using GameFrame.Gameplay;
 
 namespace GameFrame.Components
 {
-    public class Movement : Component
+    public class Movement(GameObject gameObject) : Component(gameObject)
     {
-        public Movement(GameObject gameObject) : base(gameObject)
-        {
-
-        }
-
         public override void Update()
         {
             Transform.Position += Input.GetDirection();
